Update DEPLOY_HOURLY
Added a check to see if the HOURLY_DEPLOY script is already running.   Addresses the problem of a slow update leading to multiple concurrent updates leading to an even slower update and still more concurrent updates.  This causes the Voteview service itself to become slow and ultimately fail.   DEPLOY_HOURLY now exits immediately if it finds that the DEPLOY_HOURLY script is already running.  Of course, cron will trigger another attempt to update one hour later.